Sk/test transmission (#16)
* Only logging results on t=0, don't run step(). Updated timer logic to properly decrement and move agents between states. Added additional testing

* add squashing tests

* cleanup

* changing fixed values to use pars for r0 and dur_inf

* commenting out unused code

* count alive correctly (previously was adding up 0s since logging for current timestep doesn't happen until later) & prevent gamma from returning nans by forcing scale_gamma to never be 0

* adding transmission tests

* removing lint

* removing lint

* adding notes on linting

* refactoring, moving faster_transmission into seir_abm & renaming to model.py

* linters

* beefing up tests to avoid failure due to random numbers

* linters and increasing pop size to beef up tests

* updating tests since we don't run step on day 0
