+++ src/laser_polio/archive/seir_mpm.py
