+++ tests/test_vital_dynamics.py
@@ -13,7 +13,7 @@ def setup_sim(step_size=1):
         {
             "start_date": lp.date("2020-01-01"),  # Start date of the simulation
             "dur": 30,  # Number of dur to run the simulation
-            "n_ppl": np.array([1000, 500]),  # Two nodes with populations
+            "n_ppl": np.array([10000, 5000]),  # Two nodes with populations
             "cbr": np.array([30, 25]),  # Birth rate per 1000/year
             "age_pyramid_path": "data/Nigeria_age_pyramid_2024.csv",  # From https://www.populationpyramid.net/nigeria/2024/
         }
@@ -73,6 +73,11 @@ def test_deaths_occur_step_size_1():
     # Tally deaths per day, add 1 since results are include init conditions and dur timesteps
     expected_deaths = np.bincount(dods)[: sim.pars.dur + 1]
     observed_deaths = np.sum(sim.results.deaths, axis=1)  # Sum across nodes per timestep
+    # Combine deaths on day 0 and day 1 since we don't run vital rates on day 0, we only log results
+    expected_deaths[1] += expected_deaths[0]  # Combine day 0 and day 1
+    expected_deaths = expected_deaths[1:]  # Remove day 0 since it's combined with day 1
+    observed_deaths[1] += observed_deaths[0]  # Combine day 0 and day 1
+    observed_deaths = observed_deaths[1:]  # Remove day 0 since it's combined with day 1
     assert np.array_equal(expected_deaths, observed_deaths), "Deaths did not occur on the expected days"
 
     unique_values, counts = np.unique(sim.people.disease_state[: np.sum(sim.pars.n_ppl)], return_counts=True)
@@ -94,6 +99,11 @@ def test_deaths_occur_step_size_7():
     dods = sim.people.date_of_death[: sim.people.count]  # Extract date_of_death for active individuals
     expected_deaths = np.bincount(dods, minlength=sim.pars.dur + 1)[: sim.pars.dur + 1]  # Daily expected deaths
     observed_deaths = np.sum(sim.results.deaths, axis=1)  # Daily observed deaths
+    # Combine deaths on day 0 and day 1 since we don't run vital rates on day 0, we only log results
+    expected_deaths[1] += expected_deaths[0]  # Combine day 0 and day 1
+    expected_deaths[0] = 0  # Set to 0 since nothing run this day
+    observed_deaths[1] += observed_deaths[0]  # Combine day 0 and day 1
+    observed_deaths[0] = 0  # Set to 0 since nothing run this day
     death_bins = np.digitize(np.arange(sim.pars.dur + 1), bin_edges, right=True)  # Digitize days into bins
     expected_deaths_binned = np.bincount(death_bins, weights=expected_deaths)[: len(bin_edges) - 1]  # Aggregate deaths per bin
     observed_deaths_binned = np.bincount(death_bins, weights=observed_deaths)[: len(bin_edges) - 1]  # Aggregate deaths per bin
